+++ terra_settings/views.py
@@ -1,5 +1,6 @@
 from django.conf import settings
 from mapbox_baselayer.models import MapBaseLayer
+from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework import viewsets
@@ -27,3 +28,8 @@ class SettingsView(APIView):
 class BaseLayerViewSet(viewsets.ModelViewSet):
     serializer_class = MapBaseLayerSerializer
     queryset = MapBaseLayer.objects.all()
+
+    @action(detail=True)
+    def tilejson(self, request, *args, **kwargs):
+        base_layer = self.get_object()
+        return Response(base_layer.tilejson)